requirements.txt
supabase
websockets/
requests
PyPDF2
python-docx
python-dotenv
reportlab
psycopg2-binary>=2.9.0